refactor: move buildWithDefaults to fixtures
