+++ server/apps/backend/src/app.spec.ts
@@ -2,8 +2,8 @@ import bodyParser from 'body-parser';
 import express, { Express } from 'express';
 import request from 'supertest';
 import Configuration from './config';
+import { buildWithDefaults } from './fixtures';
 import { Sellers } from './repositories';
-import { buildWithDefaults } from './repositories/Seller';
 import routes from './routes';
 import { SellerService } from './services';
