+++ server/apps/backend/src/fixtures/sellers.fixture.ts
@@ -27,3 +27,12 @@ export const cashHistory: CashHistory = {
   },
   lastIteration: 4,
 };
+
+export const buildWithDefaults = (values: Partial<Seller>): Seller => ({
+  name: 'John',
+  password: '123456',
+  url: new URL('https://192.168.0.1:3000'),
+  cash: 0,
+  online: true,
+  ...values,
+});
