+++ server/apps/backend/src/services/OrderService.spec.ts
@@ -1,8 +1,8 @@
 import { URL } from 'node:url';
 import _ from 'lodash';
 import Configuration from '../config';
+import { buildWithDefaults } from '../fixtures';
 import { Countries } from '../repositories';
-import { buildWithDefaults } from '../repositories/Seller';
 import utils from '../utils';
 import OrderService from './OrderService';
 import Reductions from './reduction';
